+++ CHANGELOG.md
@@ -37,8 +37,16 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
-- [SQS-Routing] API to force quote over a given route
+- Fix bug of overapplying quote denom scaling factor in pool liquidity pricer.
+
+## v25.1.1
+
 - Add fault tolerance in candidate route conversion when pools are breaking.
+- Hotfix for Alloyed pool swaps failing due to not having LP share in balances. 
+
+## v25.1.0
+
+- [SQS-Routing] API to force quote over a given route
 
 ## v25.0.0